bug fixes
- Fix out-of-order bug building gTest.  Added new function parameters
which are called with add_dependency, which CAN be called from utility
projects
- Added missing iostream
- Fixed a typo in the windows cross platform lib-defs
- Changed gtest prefix to be compatible with out-of-source builds (would
not build OOS otherwise based on a relative path being used)

With these changes everything builds through on the first pass.  Linking
still fails on one test because it fails to link against boost.  Good to
go.